Add files via upload
The html file contains the code for an image classifier that can train a neural network to predict flower types. An accuracy of 84% was achieved. The two py files are an application that can take user inputs for both training and prediction. The user is allowed to choose the neural network architecture, learning rate, number of epochs, the number of possible types to display and so on.